fix(gallery-web): rename and replace hooks
