+++ packages/pluggableWidgets/gallery-web/src/helpers/ClickActionHelper.ts
@@ -1,5 +1,7 @@
-import { executeAction } from "@mendix/widget-plugin-platform/framework/execute-action";
 import { ListActionValue, ObjectItem } from "mendix";
+import { useEffect, useMemo } from "react";
+import { executeAction } from "@mendix/widget-plugin-platform/framework/execute-action";
+import { GalleryContainerProps, GalleryPreviewProps } from "../../typings/GalleryProps";
 
 export type ExecuteActionFx = (item: ObjectItem) => void;
 
@@ -15,4 +17,17 @@ export class ClickActionHelper {
             executeAction(this.listAction.get(item));
         }
     };
+}
+
+export function useClickActionHelper(
+    props: Pick<GalleryContainerProps | GalleryPreviewProps, "onClick">
+): ClickActionHelper {
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    const clickActionHelper = useMemo(() => new ClickActionHelper(props.onClick), []);
+
+    useEffect(() => {
+        clickActionHelper.update(props.onClick);
+    });
+
+    return clickActionHelper;
 }
