+++ packages/pluggableWidgets/gallery-web/src/helpers/useItemSelectHelper.ts
